fastapi/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
mysql-connector-python
requests
schedule
